Lab4: finished
Moved ejs pages to the views directory because that is where ejs seems to be
looking for them. Made the server send the proper messages to the ejs templates.
Removed the minimum values on the inputs to properly test the error pages, and
added a check to data.js.

I noticed that with small inputs the retirement savings method would return NaN,
so I added an exception for this case.